refactor(admin): clean huge CandidatePool component
