+++ app/organisms/CandidatePool/FullHeightCard.tsx
@@ -1,0 +1,7 @@
+import { Card } from '@singularity/core'
+import styled from 'styled-components'
+
+export const FullHeightCard = styled(Card)`
+  padding: 0;
+  height: 100%;
+`
